Refactor routes to their own classes
