+++ backend/src/entity/Site.ts
@@ -1,6 +1,6 @@
 import {Entity, Column, PrimaryColumn, OneToMany} from 'typeorm'
 import {File} from './File'
-import {UploadedMetadata} from './UploadedMetadata'
+import {Upload} from './Upload'
 
 @Entity()
 export class Site {
@@ -32,6 +32,6 @@ export class Site {
     @OneToMany(_ => File, file => file.site)
     files!: File[]
 
-    @OneToMany(_ => UploadedMetadata, uploadedMetadata => uploadedMetadata.site)
-    uploadedMetadatas!: UploadedMetadata[]
+    @OneToMany(_ => Upload, uploadedMetadata => uploadedMetadata.site)
+    uploadedMetadatas!: Upload[]
 }
